Improved version checks
Updated URLs to work with pypi transition to warehouse
Updated version check to stop squawking on PRs
Moved version checks to separate file an deduplicated them

NOTE: Not pushing these changes to pypi as the focus is clearing the errors in builds
