+++ check_docker/check_docker.py
@@ -24,7 +24,7 @@ __author__ = 'Tim Laurence'
 __copyright__ = "Copyright 2018"
 __credits__ = ['Tim Laurence']
 __license__ = "GPL"
-__version__ = "2.0.5"
+__version__ = "2.0.6"
 
 '''
 nrpe compatible check for docker containers.
